Added Second Design
Java - WebView design is a similar design but letting the website do the Add Product logic instead of Java. This can sync up with the Web Side and make the application lighter weight. The first design would require more Java classes etc. which may slow down the application overall. Bear in mind the products tab will have to still list the products, and edit or delete each one, which makes the activity to heavy with the first design.